# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "프롬 API",
    "version": "1.0.0",
    "description": "API for creating and fetching documents"
  },
  "paths": {
...
    "/api/documents": {
      "post": {
        "summary": "Create a document",
        "tags": ["Documents CRUD"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get list of documents",
        "tags": ["Documents CRUD"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "enum": ["created_at", "updated_at", "title", "word_count"] } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "favoritesOnly", "in": "query", "schema": { "type": "boolean" } },
          { "name": "user_id", "in": "query", "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "List of documents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/documents/{id}/favorite": {
      "put": {
        "summary": "Update is_favorte",
        "tags": ["Documents management"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "is_favorite": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated is_favorite",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "Document not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/documents/recent": {
      "get": {
        "summary": "사용자의 최근 문서 조회",
        "tags": ["Recent Documents"],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "조회할 사용자의 ID"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "조회할 문서 수 (기본값: 10)"
          }
        ],
        "responses": {
          "200": {
            "description": "최근 문서 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "documents": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "example": "doc_123" },
                          "title": { "type": "string", "example": "AI Report" },
                          "content": { "type": "string", "example": "본문 내용..." },
                          "excerpt": { "type": "string", "example": "요약 내용..." },
                          "wordCount": { "type": "integer", "example": 1200 },
                          "category": { "type": "string", "example": "Work" },
                          "tags": {
                            "type": "array",
                            "items": { "type": "string" },
                            "example": ["AI", "Research"]
                          },
                          "isFavorite": { "type": "boolean", "example": false },
                          "createdAt": { "type": "string", "format": "date-time", "example": "2024-06-01T10:00:00Z" },
                          "updatedAt": { "type": "string", "format": "date-time", "example": "2024-07-01T10:00:00Z" },
                          "lastModifiedAt": { "type": "string", "format": "date-time", "example": "2024-07-15T10:00:00Z" },
                          "status": { "type": "string", "example": "draft" },
                          "aiRequestsUsed": { "type": "integer", "example": 3 },
                          "lastAccessedAt": { "type": "string", "format": "date-time", "example": "2024-07-27T12:30:00Z" },
                          "timeSpent": { "type": "number", "example": 12.5, "description": "분 단위 시간" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "필수 파라미터 누락 (user_id)"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        }
      }
    },
    "/api/documents/{id}/access": {
      "post": {
        "summary": "Create a document access log",
        "tags": ["Recent Documents"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "time_spent": {
                    "type": "integer",
                    "nullable": true
                  }
                },
                "required": ["user_id"]
              },
              "example": {
                "user_id": "a888fe3f-8a7e-4069-98e2-173c8c74c645",
                "time_spent": 120
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document access log created"
          },
          "404": {
            "description": "Document not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DocumentInput": {
        "type": "object",
        "required": ["title", "category", "user_id"],
        "properties": {
          "title": { "type": "string" },
          "content": { "type": "string" },
          "category": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "user_id": { "type": "string", "format": "uuid" }
        }
      },
      "Document": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "content": { "type": "string" },
          "excerpt": { "type": "string" },
          "word_count": { "type": "integer" },
          "category": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "string" },
          "is_favorite": { "type": "boolean" },
          "ai_requests_used": { "type": "integer" },
          "user_id": { "type": "string", "format": "uuid" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "last_modified_at": { "type": "string", "format": "date-time" }
        }
      },
      "DocumentListResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "documents": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Document" }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "total": { "type": "integer" },
              "page": { "type": "integer" },
              "limit": { "type": "integer" },
              "totalPages": { "type": "integer" }
            }
          },
          "stats": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "category": { "type": "string" },
                "_count": { "type": "object", "properties": { "_all": { "type": "integer" } } }
              }
            }
          }
        }
      }
    }
  }
}